+++ game/core/src/main/java/io/github/archessmn/ENG1/GameModel/Building.java
@@ -1,13 +1,7 @@
 package io.github.archessmn.ENG1.GameModel;
 
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Vector2;
-import io.github.archessmn.ENG1.Util.GridCoordTuple;
-import io.github.archessmn.ENG1.Util.GridUtils;
-import io.github.archessmn.ENG1.World;
 
 /**
  * Base class for each building type ({@link Building.Type}),
