+++ game/core/src/main/java/io/github/archessmn/ENG1/GameModel/GridCoordTuple.java
@@ -1,4 +1,4 @@
-package io.github.archessmn.ENG1.Util;
+package io.github.archessmn.ENG1.GameModel;
 
 /**
  * Utility class for storing the grid coordinates of a building.
