+++ game/core/src/main/java/io/github/archessmn/ENG1/GameModel/GridUtils.java
@@ -1,12 +1,10 @@
-package io.github.archessmn.ENG1.Util;
+package io.github.archessmn.ENG1.GameModel;
 
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector2;
 
-import static io.github.archessmn.ENG1.Main.VIEWPORT_HEIGHT;
-import static io.github.archessmn.ENG1.Main.VIEWPORT_WIDTH;
+import static io.github.archessmn.ENG1.Interface.Main.VIEWPORT_HEIGHT;
+import static io.github.archessmn.ENG1.Interface.Main.VIEWPORT_WIDTH;
 
 /**
  * Utilities to assist with usage of the world grid.
