+++ game/lwjgl3/src/main/java/io/github/archessmn/ENG1/lwjgl3/Lwjgl3Launcher.java
@@ -2,7 +2,7 @@ package io.github.archessmn.ENG1.lwjgl3;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-import io.github.archessmn.ENG1.Main;
+import io.github.archessmn.ENG1.Interface.Main;
 
 /** Launches the desktop (LWJGL3) application. */
 public class Lwjgl3Launcher {
